Developing Password Reset functionality
Made changes to allow for users to reset passwords. resetPassword.php has basic checking to see if the two inputs are the same, which upon confirming that the two strings are the same, will redirect the user to response.php, which is where the script to update the password will be along with notifying the user that there password has been updated.